requirements.txt
bs4
requests
python-dotenv
tgcrypto
wget
youtube-search
